complete set up of jest and enzyme -- write first front-end test of App.jsx
