+++ .babelrc
@@ -1,0 +1,6 @@
+{
+  "presets": [
+    "@babel/env",
+    "@babel/react"
+  ]
+}
