+++ .eslintrc
@@ -1,0 +1,4 @@
+
+module.exports = {
+  extends: "airbnb",
+};
