+++ .prettierrc
@@ -1,0 +1,4 @@
+{
+  "printWidth": 100,
+    "singleQuote": true
+}
