+++ STYLE-GUIDE.md
@@ -2,3 +2,14 @@
 
 Refer to the [AirBnb Style Guide](https://github.com/airbnb/javascript).
 
+
+
+{ eslint: '^4.19.1 || ^5.3.0',
+  'eslint-plugin-import': '^2.14.0',
+  'eslint-plugin-jsx-a11y': '^6.1.1',
+  'eslint-plugin-react': '^7.11.0' }
+
+
+
+
+npm install --save-dev eslint-config-airbnb eslint@^5.6.0 eslint-plugin-jsx-a11y@^6.1.1 eslint-plugin-import@^2.14.0 eslint-plugin-react@^7.11.0
