+++ spec/src/setupTests.js
@@ -1,0 +1,5 @@
+
+import { configure } from 'enzyme';
+import Adapter from 'enzyme-adapter-react-16';
+configure({ adapter: new Adapter() });
+
